+++ src/resources/hubs/invite.ts
@@ -1,6 +1,11 @@
 import * as Request from "../../common/api/request";
 import { links, StandardParams } from "../../common/api";
-import { CollectionDoc, SingleDoc, Email } from "../../common/structs";
+import {
+  CollectionDoc,
+  SingleDoc,
+  Email,
+  ResourceId,
+} from "../../common/structs";
 import * as Memberships from "./membership";
 
 export type RoleName = "owner" | "admin" | "developer" | "analyst";
@@ -10,15 +15,23 @@ export interface CreateParams {
   role: RoleName;
 }
 
+export async function remove(
+  params: StandardParams & {
+    inviteId: ResourceId;
+  },
+) {
+  return Request.deleteRequest<SingleDoc<Memberships.Membership>>({
+    ...params,
+    target: links.hubs().invites().single(params.inviteId),
+  });
+}
+
 export async function getCollection(
   params: StandardParams<Memberships.MembershipQuery>,
 ) {
   return Request.getRequest<CollectionDoc<Memberships.Membership>>({
     ...params,
-    target: links
-      .hubs()
-      .invites()
-      .collection(),
+    target: links.hubs().invites().collection(),
   });
 }
 
@@ -29,9 +42,6 @@ export async function create(
 ) {
   return Request.postRequest<SingleDoc<Memberships.Membership>>({
     ...params,
-    target: links
-      .hubs()
-      .invites()
-      .collection(),
+    target: links.hubs().invites().collection(),
   });
 }