+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Ability to delete a hub invite
 - Added `contract` to billing support plans
 - Added `infrastructure.server.services.sftp.lockdown.auto` to activity
 - Added `infrastructure.server.task.reconfigure_features` to activity